feat: 设置redis 缓存的时间

+++ server/config/cache.go
@@ -1,5 +1,5 @@
 package config
 
 type Cache struct {
-	Time int `mapstructure:"time" json:"time" yaml:"time"`
+	Time int64 `mapstructure:"time" json:"time" yaml:"time"`
 }
